moved schedule scervice to correct place
