+++ src/frontend/src/app/app.module.ts
@@ -9,7 +9,7 @@ import { routing }        from './app.routing';
 import { AuthGuard } from './guards/index';
 import { AlertComponent } from './directives/index';
 import { AlertService, AuthService, ConfigurationService, 
-  StatisticsService, AssetsService, AuditService } from './services/index';
+  StatisticsService, AssetsService, AuditService, SchedulesService } from './services/index';
 import { HomeComponent } from './home/index';
 import { LoginComponent } from './login/index';
 import { FooterComponent } from './footer/index';
@@ -58,7 +58,8 @@ import { AuditLogComponent } from '../app/audit-log/audit-log.component';
     ConfigurationService,
     StatisticsService,
     AssetsService,
-    AuditService
+    AuditService,
+    SchedulesService
   ],
   bootstrap: [ AppComponent ]
 })
