+++ src/frontend/src/app/services/statistics.service.ts
@@ -10,12 +10,6 @@ export class StatisticsService {
   private GET_STATISTICS = BASE_URL + "statistics"
   private GET_STATISTICS_HISTORY = BASE_URL + "statistics/history?limit=5"
 
-  private GET_SCHEDULE_PROCESS = BASE_URL + "schedule/process"
-  private GET_SCHEDULES = BASE_URL + "schedules"
-
-  private GET_TASKS = BASE_URL + "tasks"
-  private GET_LATEST_TASKS = BASE_URL + "tasks/latest"
-
   constructor(private http: Http) { }
 
   /**
@@ -32,33 +26,6 @@ export class StatisticsService {
    */
   public getStatisticsHistory() {
     return this.http.get(this.GET_STATISTICS_HISTORY)
-      .map(response => response.json())
-      .catch((error: Response) => Observable.throw(error.json().message || 'Server error'))
-  }
-
-  /**
-   *  GET | /foglamp/schedules 
-   */
-  public getSchedules() {
-    return this.http.get(this.GET_SCHEDULES)
-      .map(response => response.json())
-      .catch((error: Response) => Observable.throw(error.json().message || 'Server error'))
-  }
-
-  /**
-   *  GET | /foglamp/schedule/process 
-   */
-  public getScheduledProcesses() {
-    return this.http.get(this.GET_SCHEDULE_PROCESS)
-      .map(response => response.json())
-      .catch((error: Response) => Observable.throw(error.json().message || 'Server error'))
-  }
-
-  /**
-   *  GET | /foglamp/tasks/latest
-   */
-  public getLatestTasks() {
-    return this.http.get(this.GET_LATEST_TASKS)
       .map(response => response.json())
       .catch((error: Response) => Observable.throw(error.json().message || 'Server error'))
   }
